refactored code packages to com.solace.kafka.kafkaproxy
