+++ demo-consumer/src/main/java/com/solace/kafka/kafkaproxy/demo/KeyValueConsumer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.solace.kafka.wireline;
+package com.solace.kafka.kafkaproxy.demo;
 
 import org.apache.commons.cli.*;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
